Drop in Apache 2.0 license for now
We can talk more about this later if anyone would like
to use a different license. I really like Apache because it
doesn't so much constrain what others can do with the project,
but it does force them to give us credit if they reuse it,
by virtue of the NOTICE file included here that must also
be visible in their project.

Signed-off-by: Jimmy Theis <jimmy@jetheis.com>